refactor(app): move the store tests with the store and seat the orphaned helpers
Repairs from the review's bug pass, all pure moves. The eight EvidenceStore
tests and their fixtures follow the store into evidence/store.rs, so
StoreState, its byte counter and lock() return to private. serialization_failed
and the two constants only reads.rs uses move to reads.rs; SCREENSHOT_FORMAT
and REGION_DECIMAL_PLACES move to image.rs; image.rs names its one cross-seam
borrow (store::raw_sha256) explicitly; local_action.rs's doc link to
ObserverContract::prepare resolves again through a reference-style link.
